Integração com o Arduino: Troca de Chaves RSA

+++ RSA.h
@@ -11,14 +11,14 @@ class RSA
 {
     public:
 
-        char *decodifica(int *mensagemC, long long d, long long n, int quant);
-        int *codifica(char *mensagem, long long e, long long n, int quant);
-        long potencia(long long a, long long e, long long n);
-        long long mdcEstendido(long long a, long long b);
-        void divisao(long long *resto, long long *quociente, long long a, long long b);
-        long long escolheE(long long phi, long long p, long long p2, long long n);
+        char *decodifica(int *mensagemC, long d, long n, int quant);
+        int *codifica(char *mensagem, long e, long n, int quant);
+        long potencia(long a, long e, long n);
+        long mdcEstendido(long a, long b);
+        void divisao(long *resto, long *quociente, long a, long b);
+        long escolheE(long phi, long p, long p2, long n);
         long geraPrimo(long numero);
-        long long verificaPrimo(long long p);
+        long verificaPrimo(long p);
         long geraNumeroRandom();
         long geraNumeroMax(int n);
         int expModular(int a, int b, int n);
